server/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=1.0.0
pydantic>=1.10.0
httpx>=0.23.0
Jinja2>=3.1.2
SQLAlchemy==2.0.25
passlib==1.7.4
email-validator==2.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-multipart==0.0.6
alembic>=1.12,<2.0
